+++ src/main/java/project/persistence/repositories/GoalRepository.java
@@ -1,9 +1,6 @@
 package project.persistence.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import project.persistence.entities.UserExercise;
 import project.persistence.entities.UserGoal;
 
 import java.util.List;
@@ -16,5 +13,5 @@ public interface GoalRepository extends JpaRepository<UserGoal, Long> {
     
     UserGoal findOne(Long id);
 
-    List<UserGoal> findByUserID(int userID);
+    List<UserGoal> findByUserID(Long userID);
 }
