+++ src/main/java/project/service/Implementation/GoalServiceImplementation.java
@@ -2,9 +2,6 @@ package project.service.Implementation;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import project.persistence.entities.Exercise;
-import project.persistence.entities.UserExercise;
 import project.persistence.entities.UserGoal;
 import project.persistence.repositories.GoalRepository;
 import project.service.GoalService;
@@ -57,7 +54,7 @@ public class GoalServiceImplementation implements GoalService {
     }
 
     @Override
-    public List<UserGoal> findByUserID(int userID) {
+    public List<UserGoal> findByUserID(Long userID) {
         return repository.findByUserID(userID);
     }
